refactor: 카테고리 라우터-컨트롤러 분리
