+++ src/routers/search-router.js
@@ -6,11 +6,9 @@ const searchRouter = Router();
 searchRouter.get('/', async function (req, res, next) {
   try {
     // 페이지네이션
-    // url 쿼리에서 page 받기, 기본값 1
+    // url 쿼리에서 page, perPage 받기
     const page = Number(req.query.page || 1);
-    // url 쿼리에서 peRage 받기, 기본값 10
     const perPage = Number(req.query.perPage || 9);
-    // 쿼리로 keyword 변수 받아옴
     const { keyword } = req.query;
     // keyword에 맞는 상품 리스트 얻음
     const { products, total, totalPage } = await productService.searchByProductName(keyword, page, perPage);
